+++ schema.sql
@@ -13,10 +13,12 @@ CREATE TABLE `room` (
   `room_id` bigint NOT NULL AUTO_INCREMENT,
   `live_id` int DEFAULT NULL,
   `select_difficulty` int DEFAULT NULL,
+  `status` int DEFAULT NULL,
   `member1` varchar(255) DEFAULT NULL,
   `member2` varchar(255) DEFAULT NULL,
   `member3` varchar(255) DEFAULT NULL,
   `member4` varchar(255) DEFAULT NULL,
+  `owner` varchar(255) DEFAULT NULL,
   PRIMARY KEY (`room_id`),
   UNIQUE KEY `room_id` (`room_id`)
